python_model/requirements.txt
# Gold-Trade Pro - Python Dependencies v2.2
# Install with: pip install -r requirements.txt

# Core Data Processing
pandas>=1.5.0
numpy>=1.21.0

# Machine Learning
xgboost>=1.7.0
scikit-learn>=1.0.0

# Technical Analysis
ta>=0.10.0

# Financial Data
yfinance>=0.2.0

# HTTP Requests (for news APIs)
requests>=2.28.0

# Environment variable loading
python-dotenv>=0.19.0

# Scheduling (for live predictions)
schedule>=1.2.0

# Optional but recommended
matplotlib>=3.5.0    # For visualization
seaborn>=0.12.0      # For visualization
